Drop By Level/MMB removed/Ores removed
This is the update for drop data, removed mastery books because it should now drop from a level. Updated it to where you can now add a drop by level. I put in some examples for tickets. I did not have time to finish it but it should be pretty easy to implement. Ores should only be found through professions. If this repo has any issues Atnight will PM me.